* moved RmSymlink to batchCommandsMacOnly.py
